console.log: install of debian-11.2
Note the following manual step is required to install grub.

During Debian installation, the "Install the GRUB boot loader" step
will fail and manual intervention is necessary to make the installed
system bootable.
Select "Execute a shell" and do:
Exit the chroot and the shell to return to the installer and select
the "Continue without boot loader" step to finish installation.

Signed-off-by: Peter Griffin <peter.griffin@linaro.org>